==> 02_indicators.Rmd <==
---
title: "03_indicators"
output: html_notebook
---

# loading packages
```{r}
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(sjPlot))
suppressPackageStartupMessages(library(lme4))
suppressPackageStartupMessages(library(psych))
suppressPackageStartupMessages(library(matrixStats))

options(scipen = 999)
```

# reading dataset
```{r}
dat_processed <- read_csv("data/data_processed.csv")
```

# computing between-strategy variability

```{r}
dat_processed <- dat_processed %>% 
  mutate(between_strategy_sd = rowSds(as.matrix(.[21:26]))) %>% 
  select(id:regu_suppression, between_strategy_sd, everything())

dat_processed <- dat_processed %>% 
  group_by(id) %>% 
  mutate(average_between_strategy_sd = mean(between_strategy_sd, na.rm = TRUE)) %>% 
  select(id:between_strategy_sd, average_between_strategy_sd, everything())

# columns are indexed by number — might change later
```

# computing within-strategy-variability

```{r}
# for each strategy
dat_processed <- dat_processed %>%
  group_by(id) %>% 
  mutate(
    distraction_sd = sd(regu_distraction, na.rm = TRUE),
    reflection_sd = sd(regu_reflection, na.rm = TRUE),
    reappraisal_sd = sd(regu_reappraisal, na.rm = TRUE),
    rumination_sd = sd(regu_rumination, na.rm = TRUE),
    socialshare_sd = sd(regu_socialshare, na.rm = TRUE),
    suppression_sd = sd(regu_suppression, na.rm = TRUE)
  ) %>% 
  ungroup() %>%
  select(id:regu_distraction, distraction_sd, regu_reflection, reflection_sd, regu_reappraisal, reappraisal_sd, regu_rumination, rumination_sd, regu_socialshare, socialshare_sd, regu_suppression, suppression_sd, everything())

# average SD across stragegies
dat_processed <- dat_processed %>%
  mutate(average_within_strategy_sd = rowMeans(select(., c("distraction_sd", "reflection_sd", "reappraisal_sd", "rumination_sd", "socialshare_sd", "suppression_sd")))) %>% 
  select(id:average_between_strategy_sd, average_within_strategy_sd, everything())
         
# ha sok a 0, akkor mi lesz a between-strategy sd átlaggal? nagyon lecsökken
```

# computing average strategy-endorsement

```{r}
dat_processed <- dat_processed %>%
  mutate(average_beep_endorsement = rowMeans(select(., c("regu_distraction", "regu_reflection", "regu_reappraisal", "regu_rumination", "regu_socialshare", "regu_suppression"))))

dat_processed <- dat_processed %>% 
  group_by(id) %>% 
  mutate(average_overall_endorsement = mean(average_beep_endorsement, na.rm = TRUE)) %>% 
  ungroup() %>% 
  select(id:average_within_strategy_sd, average_beep_endorsement, average_overall_endorsement, everything())
```


```{r}
dat_processed %>% 
  select(id, rumination_sd) %>% 
  distinct() %>% 
  filter(rumination_sd > 0) %>% 
  ggplot(aes(x = rumination_sd)) + 
  geom_histogram()
```
